requirements.txt
absl-py==0.8.1
alembic==1.4.0
antiorm==1.2.1
astor==0.8.0
attrs==19.3.0
Babel==2.8.0
beautifulsoup4==4.4.1
blinker==1.4
bottle==0.12.18
Click==7.0
cloudpickle==1.2.2
coverage==5.0.3
cycler==0.10.0
Cython==0.29.14
darkflow==1.0.0
dask==2.6.0
db==0.1.1
db-sqlite3==0.0.1
decorator==4.4.1
Django==2.2.10
django-bootstrap3==12.0.3
django-bootstrap4==1.1.1
django-crispy-forms==1.8.1
djangorestframework==3.11.0
dlib==19.18.0
dnspython==1.16.0
elasticsearch==6.3.1
eventlet==0.25.1
face-recognition==1.2.3
face-recognition-models==0.3.0
facebook-sdk==3.1.0
facenet==1.0.5
Flask==1.0.2
Flask-BabelEx==0.9.4
Flask-Classful==0.14.2
Flask-Classy==0.6.10
Flask-Cors==3.0.8
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-Paranoid==0.2.0
Flask-Principal==0.4.0
Flask-PyMongo==2.3.0
Flask-Security==3.0.0
Flask-SQLAlchemy==2.4.1
Flask-Testing==0.7.1
Flask-WTF==0.14.3
gast==0.3.2
greenlet==0.4.15
grpcio==1.25.0
gunicorn==19.9.0
h5py==2.10.0
httplib2==0.9.1
ImageHash==4.0
importlib-metadata==1.4.0
importlib-resources==1.0.2
imutils==0.5.3
iso8601==0.1.12
itsdangerous==1.1.0
Jinja2==2.10.3
joblib==0.14.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lxml==4.4.1
Mako==1.1.1
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.0.3
MechanicalSoup==0.12.0
mock==0.8.0
monotonic==1.5
more-itertools==8.1.0
networkx==2.4
newsapi-python==0.2.6
nose==1.3.7
numpy==1.16.0
oauth2client==4.1.3
oauthlib==3.1.0
opencv-python==4.1.1.26
packaging==20.0
pandas==0.23.4
passlib==1.7.2
pathlib==1.0.1
pathlib2==2.3.5
Pillow==5.3.0
pluggy==0.13.1
protobuf==3.10.0
psutil==5.6.5
psycopg2==2.8.4
py==1.8.1
py4j==0.10.7
pyasn1==0.4.8
pyasn1-modules==0.2.7
pymongo==3.10.0
pyparsing==2.4.5
pyspark==2.4.0
pysqlite3==0.4.1
pytest==5.3.2
pytest-flask==0.15.0
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.1.2
requests==2.9.1
rsa==4.0
schedule==0.6.0
scikit-image==0.14.1
scikit-learn==0.21.3
scipy==1.3.2
simplejson==3.17.0
six==1.13.0
sklearn==0.0
speaklater==1.3
SQLAlchemy==1.3.13
sqlparse==0.3.0
tensorboard==1.9.0
tensorflow==1.9.0
termcolor==1.1.0
threaded==3.0.3
toolz==0.10.0
tqdm==4.38.0
typing==3.7.4.1
Unidecode==1.1.1
unittest-resources==0.2.0
urllib3==1.25.7
virtualenv==15.0.1
wcwidth==0.1.8
webcolors==1.10
Werkzeug==0.16.0
WTForms==2.2.1
zipp==1.0.0
